[marshal] Cleanup Validation (#3184)
